first pass with codex doing latency, esp mobile lag
