+++ package.json
@@ -6,6 +6,7 @@
     "generate:ai-surfaces": "node scripts/generate-ai-surfaces.js",
     "generate:algolia-synonyms": "node scripts/generate-algolia-synonyms.js",
     "generate:bespoke-host-pages": "node scripts/generate-bespoke-host-pages.js",
+    "generate:page-model-chunks": "node scripts/generate-page-model-chunks.js",
     "bootstrap:i18n": "node scripts/bootstrap-i18n.js",
     "bootstrap:i18n:reseed": "node scripts/bootstrap-i18n.js --reseed",
     "bootstrap:i18n:ru": "node scripts/bootstrap-ru-i18n.js",
@@ -13,12 +14,12 @@
     "submit:indexnow": "node scripts/submit-indexnow.js",
     "submit:indexnow:dry-run": "node scripts/submit-indexnow.js --dry-run",
     "docusaurus": "docusaurus",
-    "start": "yarn generate:bespoke-host-pages && docusaurus build && docusaurus serve",
-    "start:full": "yarn generate:ai-surfaces && yarn generate:bespoke-host-pages && docusaurus build && docusaurus serve",
-    "start:dev": "yarn generate:bespoke-host-pages && docusaurus start",
-    "start:dev:ru": "yarn generate:bespoke-host-pages && docusaurus start --locale ru",
-    "build": "yarn generate:ai-surfaces && yarn generate:bespoke-host-pages && docusaurus build",
-    "audit:page-models": "node scripts/audit-generated-page-models.js",
+    "start": "yarn generate:page-model-chunks && yarn generate:bespoke-host-pages && docusaurus build && docusaurus serve",
+    "start:full": "yarn generate:ai-surfaces && yarn generate:page-model-chunks && yarn generate:bespoke-host-pages && docusaurus build && docusaurus serve",
+    "start:dev": "yarn generate:page-model-chunks && yarn generate:bespoke-host-pages && docusaurus start",
+    "start:dev:ru": "yarn generate:page-model-chunks && yarn generate:bespoke-host-pages && docusaurus start --locale ru",
+    "build": "yarn generate:ai-surfaces && yarn generate:page-model-chunks && yarn generate:bespoke-host-pages && docusaurus build",
+    "audit:page-models": "node scripts/audit-generated-page-models.js && yarn generate:page-model-chunks",
     "audit:i18n": "node scripts/audit-i18n.js",
     "audit:i18n:all": "node scripts/audit-i18n.js --all-locales --wave 1",
     "audit:i18n:ru": "node scripts/audit-ru-translation.js",
@@ -36,7 +37,7 @@
     "ci:locale-quality": "yarn audit:ru-terminology && yarn audit:i18n:all && yarn build && node scripts/audit-indexing-surface.js",
     "playwright:install": "env -u NO_COLOR playwright install chromium",
     "swizzle": "docusaurus swizzle",
-    "deploy": "yarn generate:ai-surfaces && yarn generate:bespoke-host-pages && docusaurus deploy && yarn submit:indexnow",
+    "deploy": "yarn generate:ai-surfaces && yarn generate:page-model-chunks && yarn generate:bespoke-host-pages && docusaurus deploy && yarn submit:indexnow",
     "clear": "docusaurus clear",
     "serve": "docusaurus serve",
     "test:e2e": "env -u NO_COLOR playwright test",
