+++ src/data/generatedFastnearPageModelChunkManifest.json
@@ -1,0 +1,80 @@
+{
+  "fastnear-system-health": "apis/fastnear",
+  "fastnear-system-status": "apis/fastnear",
+  "fastnear-v0-account-ft": "apis/fastnear",
+  "fastnear-v0-account-nft": "apis/fastnear",
+  "fastnear-v0-account-staking": "apis/fastnear",
+  "fastnear-v0-public-key-lookup": "apis/fastnear",
+  "fastnear-v0-public-key-lookup-all": "apis/fastnear",
+  "fastnear-v1-account-ft": "apis/fastnear",
+  "fastnear-v1-account-full": "apis/fastnear",
+  "fastnear-v1-account-nft": "apis/fastnear",
+  "fastnear-v1-account-staking": "apis/fastnear",
+  "fastnear-v1-ft-top": "apis/fastnear",
+  "fastnear-v1-public-key-lookup": "apis/fastnear",
+  "fastnear-v1-public-key-lookup-all": "apis/fastnear",
+  "kv-fastdata-v0-all-by-predecessor": "apis/kv-fastdata",
+  "kv-fastdata-v0-get-history-key": "apis/kv-fastdata",
+  "kv-fastdata-v0-get-latest-key": "apis/kv-fastdata",
+  "kv-fastdata-v0-history-by-account": "apis/kv-fastdata",
+  "kv-fastdata-v0-history-by-key": "apis/kv-fastdata",
+  "kv-fastdata-v0-history-by-predecessor": "apis/kv-fastdata",
+  "kv-fastdata-v0-latest-by-account": "apis/kv-fastdata",
+  "kv-fastdata-v0-latest-by-predecessor": "apis/kv-fastdata",
+  "kv-fastdata-v0-multi": "apis/kv-fastdata",
+  "neardata-system-health": "apis/neardata",
+  "neardata-v0-block": "apis/neardata",
+  "neardata-v0-block-chunk": "apis/neardata",
+  "neardata-v0-block-headers": "apis/neardata",
+  "neardata-v0-block-optimistic": "apis/neardata",
+  "neardata-v0-block-shard": "apis/neardata",
+  "neardata-v0-first-block": "apis/neardata",
+  "neardata-v0-last-block-final": "apis/neardata",
+  "neardata-v0-last-block-optimistic": "apis/neardata",
+  "rpc-block-by-height": "rpcs/block",
+  "rpc-block-by-id": "rpcs/block",
+  "rpc-block-effects": "rpcs/block",
+  "rpc-broadcast-tx-async": "rpcs/transaction",
+  "rpc-broadcast-tx-commit": "rpcs/transaction",
+  "rpc-call-function": "rpcs/contract",
+  "rpc-changes": "rpcs/protocol",
+  "rpc-chunk-by-block-shard": "rpcs/protocol",
+  "rpc-chunk-by-hash": "rpcs/protocol",
+  "rpc-client-config": "rpcs/protocol",
+  "rpc-EXPERIMENTAL-congestion-level": "rpcs/protocol",
+  "rpc-EXPERIMENTAL-light-client-block-proof": "rpcs/protocol",
+  "rpc-EXPERIMENTAL-light-client-proof": "rpcs/protocol",
+  "rpc-EXPERIMENTAL-protocol-config": "rpcs/protocol",
+  "rpc-EXPERIMENTAL-receipt": "rpcs/transaction",
+  "rpc-EXPERIMENTAL-split-storage-info": "rpcs/protocol",
+  "rpc-EXPERIMENTAL-tx-status": "rpcs/transaction",
+  "rpc-EXPERIMENTAL-validators-ordered": "rpcs/validators",
+  "rpc-gas-price": "rpcs/protocol",
+  "rpc-gas-price-by-block": "rpcs/protocol",
+  "rpc-genesis-config": "rpcs/protocol",
+  "rpc-health": "rpcs/protocol",
+  "rpc-latest-block": "rpcs/protocol",
+  "rpc-light-client-proof": "rpcs/protocol",
+  "rpc-maintenance-windows": "rpcs/protocol",
+  "rpc-metrics": "rpcs/protocol",
+  "rpc-network-info": "rpcs/protocol",
+  "rpc-next-light-client-block": "rpcs/protocol",
+  "rpc-send-tx": "rpcs/transaction",
+  "rpc-status": "rpcs/protocol",
+  "rpc-tx-status": "rpcs/transaction",
+  "rpc-validators-by-epoch": "rpcs/validators",
+  "rpc-validators-current": "rpcs/validators",
+  "rpc-view-access-key": "rpcs/account",
+  "rpc-view-access-key-list": "rpcs/account",
+  "rpc-view-account": "rpcs/account",
+  "rpc-view-code": "rpcs/contract",
+  "rpc-view-global-contract-code": "rpcs/contract",
+  "rpc-view-global-contract-code-by-account-id": "rpcs/contract",
+  "rpc-view-state": "rpcs/contract",
+  "transactions-v0-account": "apis/transactions",
+  "transactions-v0-block": "apis/transactions",
+  "transactions-v0-blocks": "apis/transactions",
+  "transactions-v0-receipt": "apis/transactions",
+  "transactions-v0-transactions": "apis/transactions",
+  "transfers-v0-transfers": "apis/transfers"
+}
